Move canohost functions (diff licence) to separate file.
