+++ lib/misc.h
@@ -140,9 +140,6 @@ G_MODULE_EXPORT int bool2int(char *value);
 G_MODULE_EXPORT struct ns_srv_reply **srv_lookup(char *service, char *protocol, char *domain);
 G_MODULE_EXPORT void srv_free(struct ns_srv_reply **srv);
 
-G_MODULE_EXPORT char *reverse_lookup(const struct sockaddr *from_, const socklen_t fromlen_);
-G_MODULE_EXPORT void ipv64_normalise_mapped(struct sockaddr_storage *addr, socklen_t *len);
-
 G_MODULE_EXPORT char *word_wrap(const char *msg, int line_len);
 G_MODULE_EXPORT gboolean ssl_sockerr_again(void *ssl);
 G_MODULE_EXPORT int md5_verify_password(char *password, char *hash);
